Fixed typo in studentRegistrationReducer.js file, Refactored student-registaration-page component
