+++ front/src/common/redux/reducers/studentRegistrationReducer.js
